feat: make test files dry
