+++ src/test/suite/scenario1.test.ts
@@ -1,53 +1,17 @@
-import * as assert from 'assert';
 import * as vscode from 'vscode';
-import simpleGit, { CleanOptions } from 'simple-git';
-import { runRenamerScenario } from './helpers/run-renamer-scenario.function';
-import { readUpsertDiffFile } from './helpers/read-upsert-diff-file.function';
+import { genericTestScenario } from './helpers/generic-test-scenario.function';
 
-suite('Extension Test Suite', () => {
+suite('Suite Scenario 1', () => {
   vscode.window.showInformationMessage('Start all tests.');
 
-  test('Sample test', async () => {
-    const git = simpleGit({
-      baseDir:
+  test('Scenario 1', async () => {
+    await genericTestScenario({
+      projectRoot:
         '/Users/tom/Development/vscode-ext/_rename-test-spas/simple-reactive-viewmodel-example',
+      filePath: './src/app/shared/book-ui/book-list/book-list.component.html',
+      newStub: 'tom-test',
+      fileDiffPath:
+        './src/test/suite/diffs/simple-reactive-viewmodel-example.txt',
     });
-    const discardChanges = async () => {
-      await git.clean([CleanOptions.FORCE, CleanOptions.RECURSIVE]);
-      await git.checkout('.');
-    };
-
-    const notClean = await git.diff();
-    if (notClean) {
-      await discardChanges();
-    }
-
-    await runRenamerScenario(
-      '/Users/tom/Development/vscode-ext/_rename-test-spas/simple-reactive-viewmodel-example',
-      './src/app/shared/book-ui/book-list/book-list.component.html',
-      'tom-test'
-    );
-
-    const diff = await git.diff();
-
-    const fileDiff = await readUpsertDiffFile(
-      './src/test/suite/diffs/simple-reactive-viewmodel-example.txt',
-      diff
-    );
-
-    assert.strictEqual(diff, fileDiff);
-
-    await discardChanges();
   });
 });
-
-/*
-  Todo:
-
-  Make Overwrite an npm param
-
-  make tests open different projects
-
-  run all tests from a config array
-
-*/
